occupancy-sensor/requirements.txt
paho-mqtt>=1.5.1
psycopg2-binary